+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ImageOptimize Changelog
 
+## 1.4.38 - 2018.08.20
+### Changed
+* Fixed an incompatibility with the Spoon plugin by removing `matrix-field` class from field type settings
+* Fixed an erroneous 2:2 aspect ratio in the default settings
+* Moved to a modern webpack build config for the AdminCP
+* Added install confetti
+
 ## 1.4.37 - 2018.08.09
 ### Changed
 * Reverted an errant commit that removed a fix for propagating field data in multi-site environments